# Conflicts:
#	HoloApp/Assets/HoloToolkit/SpatialMapping/Materials/WireframeBlue.mat
#	HoloApp/Assets/HoloToolkit/UX/Materials/MRTK_Standard_Lime.mat
#	HoloApp/Assets/Plugins/Editor/JetBrains/JetBrains.Rider.Unity.Editor.Plugin.Repacked.dll.meta
#	HoloApp/Assets/Resources/VuforiaConfiguration.asset
#	HoloApp/Assets/Scenes/Default.unity
#	HoloApp/Assets/Vuforia/Scripts/DefaultTrackableEventHandler.cs
#	HoloApp/Assets/Vuforia/Scripts/Internal/RuntimeOpenSourceInitializer.cs
#	HoloApp/Assets/Vuforia/version
#	HoloApp/Assets/Vuforia/version.meta

+++ HoloCADExporter/HoloCADExporter/Fragment.cs
@@ -10,6 +10,6 @@ namespace WpfApplication1
         public string type;
         public double length;
         public double radius;
-        public int bendAngle;
+        public double bendAngle;
     }
 }
